requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.16
django-admin-interface==0.28.5
django-autocomplete-light==3.11.0
django-colorfield==0.11.0
django-environ==0.8.1
django-jalali==6.0.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
jdatetime==4.1.1
pillow==10.2.0
pycodestyle==2.8.0
python-slugify==8.0.4
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
